requirements.txt
diffusers==0.16.1
transformers==4.29.1
onnx==1.13.1
accelerate==0.19.0
torch==1.13.1
torchvision==0.14.1
PySimpleGUI==4.60.4
onnxruntime-directml>=1.15.0